+++ Examples/AdditionalExamples/Examples/Advanced.swift
@@ -63,7 +63,7 @@ class AdvancedViewController: UIViewController, NavigationMapViewDelegate, Navig
 
         navigationMapView = .init(
             location: mapboxNavigation.navigation()
-                .locationMatching.map(\.mapMatchingResult.enhancedLocation)
+                .locationMatching.map(\.enhancedLocation)
                 .eraseToAnyPublisher(),
             routeProgress: mapboxNavigation.navigation()
                 .routeProgress.map(\.?.routeProgress)