+++ Examples/AdditionalExamples/Examples/Beta-Query-Parameters.swift
@@ -44,7 +44,7 @@ class BetaQueryViewController: UIViewController, NavigationMapViewDelegate, Navi
 
         navigationMapView = .init(
             location: mapboxNavigation.navigation()
-                .locationMatching.map(\.mapMatchingResult.enhancedLocation)
+                .locationMatching.map(\.enhancedLocation)
                 .eraseToAnyPublisher(),
             routeProgress: mapboxNavigation.navigation()
                 .routeProgress.map(\.?.routeProgress)