+++ Examples/AdditionalExamples/Examples/Custom-Waypoints.swift
@@ -61,7 +61,7 @@ class CustomWaypointsViewController: UIViewController {
 
         navigationMapView = .init(
             location: mapboxNavigation.navigation()
-                .locationMatching.map(\.mapMatchingResult.enhancedLocation)
+                .locationMatching.map(\.enhancedLocation)
                 .eraseToAnyPublisher(),
             routeProgress: mapboxNavigation.navigation()
                 .routeProgress.map(\.?.routeProgress)