+++ Examples/AdditionalExamples/Examples/History-Replaying.swift
@@ -57,7 +57,7 @@ class HistoryReplayingViewController: UIViewController, NavigationMapViewDelegat
         if navigationMapView == nil {
             navigationMapView = .init(
                 location: mapboxNavigation.navigation()
-                    .locationMatching.map(\.mapMatchingResult.enhancedLocation)
+                    .locationMatching.map(\.enhancedLocation)
                     .eraseToAnyPublisher(),
                 routeProgress: mapboxNavigation.navigation()
                     .routeProgress.map(\.?.routeProgress)