+++ Examples/AdditionalExamples/Examples/Offline-Regions.swift
@@ -76,7 +76,7 @@ class OfflineRegionsViewController: UIViewController {
     func setupNavigationMapView() {
         let navigationMapView = NavigationMapView(
             location: mapboxNavigation.navigation()
-                .locationMatching.map(\.mapMatchingResult.enhancedLocation)
+                .locationMatching.map(\.enhancedLocation)
                 .eraseToAnyPublisher(),
             routeProgress: mapboxNavigation.navigation()
                 .routeProgress.map(\.?.routeProgress)